+++ composer.json
@@ -17,7 +17,9 @@
         }
     ],
     "require": {
-        "php": ">=8.0"
+        "php": ">=8.0",
+        "ext-simplexml": "*",
+        "ext-curl": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.4"
