+++ src/Models/Common.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LisDev\Models;
+
+class Common
+{
+
+}
