+++ src/Models/InternetDocument.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LisDev\Models;
+
+class InternetDocument
+{
+
+}
