+++ src/Services/ConnectionService.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LisDev\Services;
+
+class ConnectionService
+{
+    /**
+     * @var string Connection type (curl | file_get_contents)
+     */
+    protected string $connectionType = 'curl';
+}
